Porting scripts into the token package and updating references
